requirements.txt
Django==5.2
djangorestframework==3.16.0
asgiref==3.8.1
sqlparse==0.5.3
